requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
google-generativeai
pdfminer.six